requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Document Processing
PyMuPDF==1.23.5               # PDF processing
pytesseract==0.3.10           # OCR for scanned PDFs
pdf2image==1.16.3             # Convert PDF to images for OCR
python-docx==1.1.0            # DOCX processing
python-pptx==0.6.22           # PPTX processing
pandas==2.1.3                 # CSV processing
openpyxl==3.1.2               # Excel support

# Text Processing and NLP
langchain==0.0.335            # Text splitting and document processing
sentence-transformers==2.2.2  # Embeddings
transformers==4.35.2          # HuggingFace models
torch==2.1.1                  # PyTorch for ML models
numpy==1.24.3                 # Numerical computing

# Vector Database
chromadb==0.4.15              # Vector database

# LLM Integration
openai==1.3.5                 # OpenAI API client

# Web Framework and UI
python-multipart==0.0.6      # File upload support
aiofiles==23.2.1              # Async file handling

# Image Processing (for OCR)
Pillow==10.1.0                # Image processing

# Utilities
python-dotenv==1.0.0         # Environment variable management
tqdm==4.66.1                  # Progress bars
pathlib2==2.3.7              # Enhanced path handling

# Development and Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# System Dependencies Note:
# The following system packages may need to be installed:
# - tesseract-ocr (for OCR functionality)
# - poppler-utils (for PDF to image conversion)
#
# Ubuntu/Debian: sudo apt-get install tesseract-ocr poppler-utils
# macOS: brew install tesseract poppler
# Windows: Download and install from respective project websites